scripts/python/requirements.txt
numpy~=1.19.5
Pillow~=9.0.0
scipy~=1.6.0